Migrate `MusicCarouselShelfBasicHeader` to TS
